impliment collision mechanics, fruit eating, snake growing

+++ include/Entity.hpp
@@ -10,15 +10,15 @@
 // abstract base entity class
 class Entity {
 public:
-  Entity(double x, double y);
-  std::pair<double, double> get_pos();
+  Entity(int x, int y);
+  std::pair<int, int> get_pos();
   virtual void draw(sf::RenderWindow&) = 0;  // draw the entity
   virtual void tick(double dt) = 0;  // update the entity
   friend bool operator==(Entity &x, Entity &y); // collision detection
 
 protected:
-  void set_pos(std::pair<double, double> &pos);
-  double posX, posY;
+  void set_pos(std::pair<int, int> &pos);
+  int posX, posY;
 };
 
 #endif /* ifndef ENTITY_HPP_DEFINED */
